backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
python-jose[cryptography]
numpy
pydantic
email-validator
